Re-write the LevelCacheStore to properly support compact on-disk MerkleTree proof generation.
Implement partial tree (or trapezoid) building while generating proofs
for MerkleTrees using either the DiskStore (wasteful in terms of
storage and computation, but illustrative) or the LevelCacheStore
(optimized for storage, but does on demand partial tree building).

Add unit test that shows the flow of MT creation using a DiskStore,
then compaction and then later LevelCacheStore usage with proof
generation.